Streamlit_project_jedha/requirements.txt
python-3.13.9
streamlit-1.54.0
pandas-3.0.0
numpy-2.4.2
matplotlib-3.10.8
seaborn-0.13.2
plotly-6.5.2